Update patch set 2
Patch Set 2:

Why would you remove the assert?  It's better to fail if the recovery.img is too big.  This call is in if !BOARD_CUSTOM_BOOTIMG_MK so it won't get called if it is not in the BOARD_CUSTOM_BOOTIMG_MK. 

https://github.com/CyanogenMod/android_build/blob/cm-13.0/core/Makefile#L972

Patch-set: 2
Reviewer: David Hacker <8640@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0