Update patch set 2
Patch Set 2:

David, for some reason the line you linked is executed (and this is why we got confused).

We still need the assert because we are creating a new image to include the device tree blob, but I don't think this should happen. This could be problematic in case the purpose of the custom script is to produce smaller images by using different compression algorithms.

Patch-set: 2