Update patch set 1
Patch Set 1:

What I meant with "right" is that this doesn't seem to be what allows to select 3G for the 2nd SIM. This should simply set the default network type, as the name suggests. I couldn't find anything that depends on this for the network selection.

Patch-set: 1
Label: Code-Review=0